[color-picker-form-component] add color input on Form Example
