+++ src/components/forms/FormInputColor/FormInputColor_Form.tsx
@@ -48,7 +48,7 @@ export const FormInputColor_Form: React.FC<FormInputColorProps> = (formProps) =>
   )
 
   return (
-    <form onSubmit={onSubmit} className='max-w-4xl md:w-full'>
+    <>
       <div className='flex justify-end'>
         <select
           className='mb-4 select select-bordered'
@@ -65,51 +65,53 @@ export const FormInputColor_Form: React.FC<FormInputColorProps> = (formProps) =>
           ))}
         </select>
       </div>
-      <div className='hidden sm:block' aria-hidden='true'>
-        <div className='py-5'>
-          <div className='border-t' />
-        </div>
-      </div>
-      <div>
-        <div className='md:grid md:grid-cols-3 md:gap-6'>
-          <div className='md:col-span-1'>
-            <div className='px-4 sm:px-0'>
-              <h3 className='text-lg font-medium leading-6 text-primary'>Form Example</h3>
-              <h3 className='my-2 text-sm font-medium leading-4 text-secondary'>color_input</h3>
-            </div>
+      <form onSubmit={onSubmit} className='max-w-4xl md:w-full'>
+        <div className='hidden sm:block' aria-hidden='true'>
+          <div className='py-5'>
+            <div className='border-t' />
           </div>
-          <div className='mt-5 md:mt-0 md:col-span-2'>
-            <div className='shadow sm:rounded-md sm:overflow-hidden'>
-              <div className='px-4 py-5 space-y-6 sm:p-6'>
-                <FormInputColor
-                  {...formProps}
-                  register={register}
-                  watch={watch}
-                  setValue={setValue}
-                  validationErrors={validationErrors}
-                />
+        </div>
+        <div>
+          <div className='md:grid md:grid-cols-3 md:gap-6'>
+            <div className='md:col-span-1'>
+              <div className='px-4 sm:px-0'>
+                <h3 className='text-lg font-medium leading-6 text-primary'>Form Example</h3>
+                <h3 className='my-2 text-sm font-medium leading-4 text-secondary'>color_input</h3>
+              </div>
+            </div>
+            <div className='mt-5 md:mt-0 md:col-span-2'>
+              <div className='shadow sm:rounded-md sm:overflow-hidden'>
+                <div className='px-4 py-5 space-y-6 sm:p-6'>
+                  <FormInputColor
+                    {...formProps}
+                    register={register}
+                    watch={watch}
+                    setValue={setValue}
+                    validationErrors={validationErrors}
+                  />
 
-                <div className='flex flex-col'>
-                  <div className='flex justify-end'>
-                    <button type='reset' className='mx-3 btn btn-secondary'>
-                      RESET
-                    </button>
+                  <div className='flex flex-col'>
+                    <div className='flex justify-end'>
+                      <button type='reset' className='mx-3 btn btn-secondary'>
+                        RESET
+                      </button>
 
-                    <button type='submit' className='mx-3 btn btn-primary'>
-                      SEND
-                    </button>
+                      <button type='submit' className='mx-3 btn btn-primary'>
+                        SEND
+                      </button>
+                    </div>
                   </div>
                 </div>
               </div>
             </div>
           </div>
         </div>
-      </div>
-      <div className='hidden sm:block' aria-hidden='true'>
-        <div className='py-5'>
-          <div className='border-t ' />
+        <div className='hidden sm:block' aria-hidden='true'>
+          <div className='py-5'>
+            <div className='border-t ' />
+          </div>
         </div>
-      </div>
-    </form>
+      </form>
+    </>
   )
 }
