+++ src/pages/form-example/form-example-filled.tsx
@@ -1,8 +1,12 @@
 import Head from 'next/head'
 import * as React from 'react'
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 import { FormExample } from '../../components/FormExample/FormExample'
+import { CodeBlock } from '../../components/forms/CodeBlock/CodeBlock'
 import { Layout } from '../../components/Layout/Layout'
 import { LinksList } from '../../model/site/LinksList'
+import { checkFetchJsonResult } from '../../utils/checkFetchResult'
 
 const Page: React.FunctionComponent = () => {
   return (
@@ -27,6 +31,28 @@ const Page: React.FunctionComponent = () => {
               color_select: 'red',
               toggle: true,
               image_url: 'https://via.placeholder.com/1080x1920.png?text=Image+Placeholder',
+              color_input: '#ff0000',
+            }}
+            onSubmitConfirm={async (submitProps) => {
+              const headers = new Headers()
+              headers.append('Content-Type', 'application/json')
+              const fetchResponse = await fetch('/api/formExample_api', {
+                method: 'POST',
+                headers,
+                body: JSON.stringify(submitProps),
+              })
+
+              /* check for server errors (VALIDATIONS) */
+              const isValid = await checkFetchJsonResult(fetchResponse)
+              if (isValid) {
+                // const resultJSON = await fetchResponse.json()
+                const myAlert = withReactContent(Swal)
+                await myAlert.fire({
+                  title: 'submited',
+                  html: <CodeBlock content={submitProps} />,
+                  confirmButtonText: 'close',
+                })
+              }
             }}
           />
         </main>
